feat:完善部分接口

+++ console/src/main/java/net/somta/juggle/console/service/IVariableInfoService.java
@@ -29,7 +29,12 @@ public interface IVariableInfoService {
      */
     Boolean updateVariable(VariableInfo variableInfo);
 
-    List<Variable> getFlowVariableList(Integer flowDefinitionId);
+    /**
+     * 获取流程变量列表
+     * @param flowDefinitionId
+     * @return
+     */
+    List<Variable> getFlowVariableList(Long flowDefinitionId);
